chore: 프로젝트 디렉토리 구조 변경
