+++ LCH-1228/nbcampContactsApp/Model/ImageData.swift
