+++ LCH-1228/nbcampContactsApp/View/ContactsListCell.swift
